// src/lib/supabase/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
// ...
      profiles: {
        Row: {
          email: string | null
          first_name: string | null
          id: string
          last_name: string | null
          notification_email: string | null
          notification_frequency:
            | Database["public"]["Enums"]["notification_frequency"]
            | null
          notification_minimum_amount: number | null
          subscription_end_date: string | null
          subscription_status: string | null
          subscription_tier:
            | Database["public"]["Enums"]["subscription_tier"]
            | null
        }
        Insert: {
          email?: string | null
          first_name?: string | null
          id: string
          last_name?: string | null
          notification_email?: string | null
          notification_frequency?:
            | Database["public"]["Enums"]["notification_frequency"]
            | null
          notification_minimum_amount?: number | null
          subscription_end_date?: string | null
          subscription_status?: string | null
          subscription_tier?:
            | Database["public"]["Enums"]["subscription_tier"]
            | null
        }
        Update: {
          email?: string | null
          first_name?: string | null
          id?: string
          last_name?: string | null
          notification_email?: string | null
          notification_frequency?:
            | Database["public"]["Enums"]["notification_frequency"]
            | null
          notification_minimum_amount?: number | null
          subscription_end_date?: string | null
          subscription_status?: string | null
          subscription_tier?:
            | Database["public"]["Enums"]["subscription_tier"]
            | null
        }
        Relationships: []
      }
      purchases: {
        Row: {
          created_at: string | null
          email: string
          expiration_date: string | null
          id: string
          metadata: Json | null
          product_type: Database["public"]["Enums"]["product_type"]
          status: string | null
          stripe_customer_id: string | null
          stripe_subscription_id: string | null
          tier: Database["public"]["Enums"]["subscription_tier"] | null
          updated_at: string | null
          user_id: string | null
        }
        Insert: {
          created_at?: string | null
          email: string
          expiration_date?: string | null
          id?: string
          metadata?: Json | null
          product_type: Database["public"]["Enums"]["product_type"]
          status?: string | null
          stripe_customer_id?: string | null
          stripe_subscription_id?: string | null
          tier?: Database["public"]["Enums"]["subscription_tier"] | null
          updated_at?: string | null
          user_id?: string | null
        }
        Update: {
          created_at?: string | null
          email?: string
          expiration_date?: string | null
          id?: string
          metadata?: Json | null
          product_type?: Database["public"]["Enums"]["product_type"]
          status?: string | null
          stripe_customer_id?: string | null
          stripe_subscription_id?: string | null
          tier?: Database["public"]["Enums"]["subscription_tier"] | null
          updated_at?: string | null
          user_id?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      domain_stats: {
        Row: {
          avg_offer: number | null
          domain: string | null
          last_offer: string | null
          offer_count: number | null
          top_offer: number | null
          visits: number | null
        }
        Relationships: []
      }
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      notification_frequency:
        | "daily"
        | "weekly"
        | "on_demand"
        | "never"
        | "monthly"
      product_type: "subscription" | "template" | "other"
      subscription_tier: "basic" | "premium" | "professional" | "template"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never